+++ components/time_formatter.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "common.h"
 
-string format_display_time(float total_seconds);
+string format_display_time(double total_seconds);
 
 // Parses an LRC timestamp like "01:23.45" (mm:ss.cc, no brackets) into total
 // seconds. Returns -1 for anything that isn't a valid mm:ss(.cc) timestamp
